feat(ui-tars-sdk): support histroy context when agent running
