+++ apps/ui-tars/src/renderer/src/hooks/useRunAgent.ts
@@ -11,16 +11,58 @@ import { getState } from '@renderer/hooks/useStore';
 import { usePermissions } from './usePermissions';
 import { useSetting } from './useSetting';
 import { api } from '@renderer/api';
-import { useSessionStore } from '../store/session';
+import { ConversationWithSoM } from '@/main/shared/types';
+import { Message } from '@ui-tars/shared/types';
+
+const filterAndTransformWithMap = (
+  history: ConversationWithSoM[],
+): Message[] => {
+  return history
+    .map((conv) => {
+      if (conv.from === 'human' && conv.value && conv.value !== '<image>') {
+        return {
+          from: conv.from,
+          value: conv.value,
+        };
+      } else if (conv.from === 'gpt' && conv.predictionParsed?.length) {
+        const finished = conv.predictionParsed.find(
+          (p) => p.action_type === 'finished' && p.action_inputs?.content,
+        );
+        if (finished) {
+          return {
+            from: conv.from,
+            value: finished.action_inputs!.content!,
+          };
+        }
+
+        const callUser = conv.predictionParsed.find(
+          (p) => p.action_type === 'call_user' && p.thought,
+        );
+        if (callUser) {
+          return {
+            from: conv.from,
+            value: callUser.thought!,
+          };
+        }
+        return undefined;
+      } else {
+        return undefined;
+      }
+    })
+    .filter((msg): msg is Message => msg !== undefined);
+};
 
 export const useRunAgent = () => {
   // const dispatch = useDispatch();
   const toast = useToast();
   const { settings } = useSetting();
   const { ensurePermissions } = usePermissions();
-  const sessionStore = useSessionStore();
 
-  const run = async (value: string, callback: () => void = () => {}) => {
+  const run = async (
+    value: string,
+    history: ConversationWithSoM[],
+    callback: () => void = () => {},
+  ) => {
     if (
       !ensurePermissions?.accessibility ||
       !ensurePermissions?.screenCapture
@@ -68,10 +110,16 @@ export const useRunAgent = () => {
     const currentMessages = getState().messages;
     console.log('initialMessages', initialMessages, currentMessages.length);
 
+    const sessionHistory = filterAndTransformWithMap(history);
+
+    console.log('sessionHistory', sessionHistory);
+
     await Promise.all([
       api.setInstructions({ instructions: value }),
       api.setMessages({ messages: [...currentMessages, ...initialMessages] }),
-      api.setSessionHistoryMessages({ messages: sessionStore.chatMessages }),
+      api.setSessionHistoryMessages({
+        messages: sessionHistory,
+      }),
     ]);
 
     await api.runAgent();
