+++ packages/ui-tars/sdk/README.md
@@ -112,7 +112,7 @@ Basic usage is largely derived from package `@ui-tars/sdk`, here's a basic examp
 > - Screenshot capture
 
 ```ts
-import { GUIAgent, UITarsModelVersion } from '@ui-tars/sdk';
+import { GUIAgent } from '@ui-tars/sdk';
 import { NutJSOperator } from '@ui-tars/operator-nut-js';
 
 const guiAgent = new GUIAgent({
@@ -128,7 +128,6 @@ const guiAgent = new GUIAgent({
   onError: ({ data, error }) => {
     console.error(error, data);
   },
-  uiTarsVersion: UITarsModelVersion.V1_5,
 });
 
 await guiAgent.run('send "hello world" to x.com');
